Pass all first chance exceptions on Windows
This changes the exception handling logic on Windows to always give the
debuggee a chance to handle all first chance exceptions (previously
STATUS_BREAKPOINT exceptions were special cased to always break in to
the debugger).

Allowing the inferior to handle breakpoint exceptions in the first
chance fixes debugging programs that use traps to redirect control flow
in structured exception handlers. Such programs tend to exhibit
different behaviour depending on whether the exception handler receives
the exception record on the first or second chance.